requirements.txt
annotated-types==0.7.0
anyio==3.7.1
bcrypt==4.0.1
certifi==2026.4.22
charset-normalizer==3.4.7
click==8.3.3
exceptiongroup==1.3.1
fastapi==0.104.1
h11==0.16.0
idna==3.13
mysql-connector-python==8.2.0
passlib==1.7.4
protobuf==4.21.12
pydantic==2.5.0
pydantic_core==2.14.1
python-multipart==0.0.6
requests==2.31.0
sniffio==1.3.1
starlette==0.27.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.7.0
uvicorn==0.24.0
